Re-structed the classes related to routing in to one place
