+++ Neo4j.Driver/Neo4j.Driver/Internal/Routing/RoundRobinClusterView.cs
@@ -18,7 +18,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace Neo4j.Driver.Internal
+namespace Neo4j.Driver.Internal.Routing
 {
     internal class RoundRobinClusterView
     {
