+++ Neo4j.Driver/Neo4j.Driver/Internal/RoutingDriver.cs
@@ -16,6 +16,7 @@
 // limitations under the License.
 
 using System;
+using Neo4j.Driver.Internal.Routing;
 using Neo4j.Driver.V1;
 
 namespace Neo4j.Driver.Internal
